requirements.txt
attrs==25.3.0
behave==1.3.1
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
locust==2.38.1
locust-cloud==1.26.3
MarkupSafe==3.0.2
msgpack==1.1.1
parse==1.20.2
parse_type==0.6.6
platformdirs==4.3.8
psutil==7.0.0
pycparser==2.22
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pywin32==311
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
tenacity==9.1.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
